requirements.txt
confluent-kafka[avro]==1.4.0
avro-python3==1.8.2
PyYAML==5.3.1